requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
phonenumbers==8.12.55
pytz==2022.2.1
sqlparse==0.4.2
